refactor(VehicleRentalOverlay): Cleanup DefaultMarkers and other changes from PR comments.
